requirements.txt
asgiref==3.3.1
awesomeversion==21.2.3
beautifulsoup4
browserstack-local
cryptography==3.3.2
celery==5.1.2
dj-database-url==0.5.0
Django==3.1.7
django-filter
django-cors-headers==3.7.0
django-extensions==3.0.8
django-js-asset==1.2.2
djangorestframework
django-rest-auth[with_social]
django-datatables-view
django-crontab
django_cron
django-celery-beat==2.2.1
django-celery-results==2.2.0
gunicorn==20.1.0
gevent
jsbeautifier
Pillow
pyimgur
psycopg2-binary==2.8.6
pytz==2021.1
PyYAML==5.4.1
requests
redis==4.1.4
selenium
whitenoise
webdriver-manager
Unidecode
undetected-chromedriver

